+++ R/PhyloDistance.R
@@ -73,7 +73,7 @@ JRFDist <- function(val, RawScore=FALSE){
   }
   if (val[1] == 0) return(1)
   maxval <- (val[2] + val[3])
-  retval <- 1 - (val[1] / maxval)
+  retval <- val[1] / maxval
   if (maxval == 0)
     retval <- as.integer(val[1] != 0)
   return(retval)